# openapi.yaml
openapi: 3.1.0
info:
  title: Hamlib REST API Server
  description: REST API for Hamlib rig control
  version: '1.1'
paths:
  /login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: login_login_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBasic: []
  /devices/usb:
    get:
      tags:
      - System
      summary: List USB serial devices
      operationId: list_usb_devices_devices_usb_get
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /hamlib/models:
    get:
      tags:
      - Hamlib
      summary: List known Hamlib rig models
      operationId: list_hamlib_models_hamlib_models_get
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /hamlib/constants:
    get:
      tags:
      - Hamlib
      summary: Enumerate Hamlib constant groups
      operationId: list_hamlib_constants_hamlib_constants_get
      parameters:
      - description: Optional group filter
        required: false
        schema:
          type: string
          title: Group
          description: Optional group filter
        name: group
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/init:
    post:
      tags:
      - Rig
      summary: Initialize rig connection
      operationId: rig_init_rig_init_post
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitRigRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/close:
    post:
      tags:
      - Rig
      summary: Close rig connection
      operationId: rig_close_rig_close_post
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/info:
    get:
      tags:
      - Rig
      summary: Get backend rig info string
      operationId: get_rig_info_rig_info_get
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/capabilities:
    get:
      tags:
      - Rig
      summary: Get rig capability bitmasks decoded
      operationId: get_rig_capabilities_rig_capabilities_get
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/frequency:
    get:
      tags:
      - Frequency
      summary: Get Freq
      operationId: get_freq_rig_frequency_get
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Frequency
      summary: Set Freq
      operationId: set_freq_rig_frequency_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Frequency'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/vfo:
    get:
      tags:
      - VFO
      summary: Get Vfo
      operationId: get_vfo_rig_vfo_get
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - VFO
      summary: Set Vfo
      operationId: set_vfo_rig_vfo_post
      parameters:
      - required: true
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/mode:
    get:
      tags:
      - Mode
      summary: Get Mode
      operationId: get_mode_rig_mode_get
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Mode
      summary: Set Mode
      operationId: set_mode_rig_mode_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mode'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/passband:
    get:
      tags:
      - Mode
      summary: Get Passband
      operationId: get_passband_rig_passband_get
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Mode
      summary: Set Passband
      operationId: set_passband_rig_passband_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Passband'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/level:
    get:
      tags:
      - Levels
      summary: Get Level
      operationId: get_level_rig_level_get
      parameters:
      - required: true
        schema:
          type: string
          title: Level
        name: level
        in: query
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Levels
      summary: Set Level
      operationId: set_level_rig_level_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelChange'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/function:
    get:
      tags:
      - Functions
      summary: Get Function
      operationId: get_function_rig_function_get
      parameters:
      - required: true
        schema:
          type: string
          title: Function
        name: function
        in: query
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Functions
      summary: Set Function
      operationId: set_function_rig_function_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionState'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/parameter:
    get:
      tags:
      - Parameters
      summary: Get Parameter
      operationId: get_parameter_rig_parameter_get
      parameters:
      - required: true
        schema:
          type: string
          title: Parameter
        name: parameter
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Parameters
      summary: Set Parameter
      operationId: set_parameter_rig_parameter_post
      parameters:
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParameterState'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rig/split:
    get:
      tags:
      - Split
      summary: Get Split
      operationId: get_split_rig_split_get
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Split
      summary: Set Split
      operationId: set_split_rig_split_post
      parameters:
      - required: false
        schema:
          type: string
          title: Vfo
        name: vfo
        in: query
      - required: true
        schema:
          type: string
          title: Session Id
        name: session_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitConfig'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Frequency:
      properties:
        frequency:
          type: number
          title: Frequency
          description: Frequency in Hz
      type: object
      required:
      - frequency
      title: Frequency
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InitRigRequest:
      properties:
        model:
          type: integer
          title: Model
          description: Hamlib rig model number
        port:
          type: string
          title: Port
          description: Serial/USB/TCP rig path
        baud:
          type: integer
          title: Baud
          description: Serial baud rate
        conf:
          type: object
          title: Conf
          description: Optional backend config parameters to apply on init
      type: object
      required:
      - model
      title: InitRigRequest
...
    Mode:
      properties:
        mode:
          type: string
          title: Mode
          description: Mode name, e.g. USB or RIG_MODE_USB
        width:
          type: integer
          title: Width
          description: Filter width / passband in Hz
      type: object
      required:
      - mode
      title: Mode
...
    Passband:
      properties:
        width:
          type: integer
          title: Width
          description: Passband width in Hz
      type: object
      required:
      - width
      title: Passband
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic